bot/requirements.txt
Pillow
pytesseract
python-telegram-bot
